requirements.txt
fastapi
uvicorn
requests
python-multipart
itsdangerous
python-dotenv
requests
sqlalchemy
pydantic[email]
